Switch to `table!()` based schema definition
We cannot use `infer_schema!()` with enum type right now, so we
need to switch for that reason anyway. But apart from that, the
`infer_schema!()` macro was a bit *too* magical for my taste anyway.
Defining the schema via `table!()` manually is a bit more work and
leads to duplication (definition in DB and in code), but I think
it's a good choice.